move Widget code to separate module
